+++ data/plugins/Hello/hello_world.py
@@ -1,0 +1,4 @@
+def main(window, document, settings):
+    window.warning(
+        settings["greeting"]
+    )