+++ src/scripting/python/python_engine.cpp
@@ -8,6 +8,7 @@
 
 scripting::ScriptEngine::Autoregister<scripting::python::PythonEngine> scripting::python::PythonEngine::autoreg;
 
+static int counter = 0;
 
 class scripting::python::PythonContext::Private
 {
@@ -15,24 +16,33 @@ public:
     pybind11::module my_module;
     py::dict globals;
     py::function compile;
+    const ScriptEngine* engine;
 };
 
-scripting::python::PythonContext::PythonContext()
+scripting::python::PythonContext::PythonContext(const ScriptEngine* engine)
 {
-    py::initialize_interpreter();
+    /// @note Not thread safe, pybind11 doesn't support multiple interpreters at once
+    if ( counter == 0 )
+        py::initialize_interpreter();
+    counter++;
+
     d = std::make_unique<Private>();
     d->my_module = py::module::import("glaxnimate");
     d->globals = py::globals();
     d->compile = py::function(py::module(d->globals["__builtins__"]).attr("compile"));
+    d->engine = engine;
 }
 
 scripting::python::PythonContext::~PythonContext()
 {
     d.reset();
-    py::finalize_interpreter();
+
+    counter--;
+    if ( counter == 0 )
+        py::finalize_interpreter();
 }
 
-void scripting::python::PythonContext::expose(const QString& name, QObject* obj)
+void scripting::python::PythonContext::expose(const QString& name, const QVariant& obj)
 {
     d->globals[name.toStdString().c_str()] = obj;
 }
@@ -59,6 +69,58 @@ QString scripting::python::PythonContext::eval_to_string(const QString& code)
     }
 }
 
+class ModuleSetter
+{
+public:
+    ModuleSetter(const QString& append)
+    {
+        sys = py::module::import("sys");
+        python_path = py::list(sys.attr("path"));
+        py::list python_path_new(sys.attr("path"));
+        python_path_new.append(append);
+        py::setattr(sys, "path", python_path_new);
+    }
+
+    ~ModuleSetter()
+    {
+        py::setattr(sys, "path", python_path);
+    }
+
+    py::module sys;
+    py::list python_path;
+};
+
+bool scripting::python::PythonContext::run_from_module (
+    const QDir& path,
+    const QString& module,
+    const QString& function,
+    const QVariantList& args
+)
+{
+    ModuleSetter{path.path()};
+
+    try {
+        py::module exec_module = py::module::import(module.toStdString().c_str());
+        std::string std_func = function.toStdString();
+        if ( !py::hasattr(exec_module, std_func.c_str()) )
+            return false;
+
+        py::tuple py_args(args.size());
+        int i = 0;
+        for ( const auto& arg : args )
+            py_args[i++] = arg;
+        exec_module.attr(std_func.c_str())(*py_args);
+    } catch ( const py::error_already_set& pyexc ) {
+        throw ScriptError(pyexc.what());
+    }
+
+    return true;
+}
+
+const scripting::ScriptEngine * scripting::python::PythonContext::engine() const
+{
+    return d->engine;
+}
 
 PYBIND11_EMBEDDED_MODULE(glaxnimate, m)
 {