+++ src/scripting/python/python_engine.hpp
@@ -7,12 +7,23 @@ namespace scripting::python {
 class PythonContext : public ScriptExecutionContext
 {
 public:
-    PythonContext();
+    PythonContext(const ScriptEngine* engine);
     ~PythonContext();
 
-    void expose(const QString& name, QObject* obj);
+    void expose(const QString& name, const QVariant& obj) override;
 
-    QString eval_to_string(const QString& code);
+    QString eval_to_string(const QString& code) override;
+
+    bool run_from_module (
+        const QDir& path,
+        const QString& module,
+        const QString& function,
+        const QVariantList& args
+    ) override;
+
+
+    const ScriptEngine* engine() const override;
+
 
 private:
     class Private;
@@ -28,7 +39,7 @@ public:
 
     ScriptContext create_context() const override
     {
-        return std::make_unique<PythonContext>();
+        return std::make_unique<PythonContext>(this);
     }
 
 private: