+++ CHANGELOG.md
@@ -2,3 +2,5 @@
  * Layer Management
  * Open/Save Glaxnim files
  * JS Scripting capabilities
+ * Python Scripting capabilities
+ * Menu action plugins